Added file opening through command line arguments
The first command argument passed to the program is used to read a file
into the editor. If successful, it sets the document name to the
provided argument. Otherwise, the document is empty and titled
"Untitled"
Also updated the OPENFILE command to set the document title to the
opened file.